+++ src/accessory/DehumidifierAccessory.ts
@@ -1,5 +1,3 @@
-import { TuyaDeviceSchemaIntegerProperty } from '../device/TuyaDevice';
-import { limit } from '../util/util';
 import BaseAccessory from './BaseAccessory';
 import { configureActive } from './characteristic/Active';
 import { configureCurrentTemperature } from './characteristic/CurrentTemperature';
@@ -7,6 +5,7 @@ import { configureCurrentRelativeHumidity } from './characteristic/CurrentRelati
 import { configureRotationSpeedLevel } from './characteristic/RotationSpeed';
 import { configureSwingMode } from './characteristic/SwingMode';
 import { configureLockPhysicalControls } from './characteristic/LockPhysicalControls';
+import { configureRelativeHumidityDehumidifierThreshold } from './characteristic/RelativeHumidityDehumidifierThreshold';
 
 const SCHEMA_CODE = {
   ACTIVE: ['switch'],
@@ -33,7 +32,7 @@ export default class DehumidifierAccessory extends BaseAccessory {
 
     // Optional Characteristics
     configureLockPhysicalControls(this, this.mainService(), this.getSchema(...SCHEMA_CODE.LOCK));
-    this.configureRelativeHumidityDehumidifierThreshold();
+    configureRelativeHumidityDehumidifierThreshold(this, this.mainService(), this.getSchema(...SCHEMA_CODE.TARGET_HUMIDITY));
     configureRotationSpeedLevel(this, this.mainService(), this.getSchema(...SCHEMA_CODE.SPEED_LEVEL));
     configureSwingMode(this, this.mainService(), this.getSchema(...SCHEMA_CODE.SWING));
 
@@ -71,33 +70,6 @@ export default class DehumidifierAccessory extends BaseAccessory {
       .onGet(() => {
         return DEHUMIDIFIER;
       }).setProps({ validValues });
-  }
-
-  configureRelativeHumidityDehumidifierThreshold() {
-    const schema = this.getSchema(...SCHEMA_CODE.TARGET_HUMIDITY);
-    if (!schema) {
-      this.log.warn('RelativeHumidityDehumidifierThreshold not supported.');
-      return;
-    }
-
-    const property = schema.property as TuyaDeviceSchemaIntegerProperty;
-    const multiple = Math.pow(10, property.scale);
-    const props = {
-      minValue: 0,
-      maxValue: 100,
-      minStep: Math.max(1, property.step / multiple),
-    };
-    this.log.debug('Set props for RelativeHumidityDehumidifierThreshold:', props);
-
-    this.mainService().getCharacteristic(this.Characteristic.RelativeHumidityDehumidifierThreshold)
-      .onGet(() => {
-        const status = this.getStatus(schema.code)!;
-        return limit(status.value as number / multiple, 0, 100);
-      })
-      .onSet(value => {
-        const dehumidity_set = limit(value as number * multiple, property.min, property.max);
-        this.sendCommands([{ code: schema.code, value: dehumidity_set }]);
-      }).setProps(props);
   }
 
 }