+++ src/accessory/characteristic/RotationSpeed.ts
@@ -3,7 +3,12 @@ import { TuyaDeviceSchema, TuyaDeviceSchemaEnumProperty, TuyaDeviceSchemaInteger
 import { limit, remap } from '../../util/util';
 import BaseAccessory from '../BaseAccessory';
 
-export function configureRotationSpeed(accessory: BaseAccessory, service: Service, schema?: TuyaDeviceSchema) {
+export function configureRotationSpeed(
+  accessory: BaseAccessory,
+  service: Service,
+  schema?: TuyaDeviceSchema,
+) {
+
   if (!schema) {
     return;
   }
@@ -22,36 +27,55 @@ export function configureRotationSpeed(accessory: BaseAccessory, service: Servic
     });
 }
 
-export function configureRotationSpeedLevel(accessory: BaseAccessory, service: Service, schema?: TuyaDeviceSchema) {
+export function configureRotationSpeedLevel(
+  accessory: BaseAccessory,
+  service: Service,
+  schema?: TuyaDeviceSchema,
+  ignoreValues?: string[],
+) {
+
   if (!schema) {
     return;
   }
 
   const property = schema.property as TuyaDeviceSchemaEnumProperty;
+  const range: string[] = [];
+  for (const value of property.range) {
+    if (ignoreValues?.includes(value)) {
+      continue;
+    }
+    range.push(value);
+  }
+
   const props = { minValue: 0, maxValue: 100, minStep: 1 };
-  props.minStep = Math.floor(100 / property.range.length);
-  props.maxValue = props.minStep * property.range.length;
+  props.minStep = Math.floor(100 / range.length);
+  props.maxValue = props.minStep * range.length;
   accessory.log.debug('Set props for RotationSpeed:', props);
 
   service.getCharacteristic(accessory.Characteristic.RotationSpeed)
     .onGet(() => {
       const status = accessory.getStatus(schema.code)!;
-      const index = property.range.indexOf(status.value as string);
+      const index = range.indexOf(status.value as string);
       return props.minStep * (index + 1);
     })
     .onSet(value => {
       const index = value as number / props.minStep - 1;
-      if (index < 0) {
+      if (index < 0 || index >= range.length) {
         return;
       }
-      value = property.range[index].toString();
-      accessory.log.debug('Set RotationSpeed to:', value);
-      accessory.sendCommands([{ code: schema.code, value }], true);
+      const speedLevel = range[index].toString();
+      accessory.log.debug('Set RotationSpeed to:', speedLevel);
+      accessory.sendCommands([{ code: schema.code, value: speedLevel }], true);
     })
     .setProps(props);
 }
 
-export function configureRotationSpeedOn(accessory: BaseAccessory, service: Service, schema?: TuyaDeviceSchema) {
+export function configureRotationSpeedOn(
+  accessory: BaseAccessory,
+  service: Service,
+  schema?: TuyaDeviceSchema,
+) {
+
   if (!schema) {
     return;
   }